requirements.txt
streamlit
ultralytics
opencv-python-headless
Pillow
streamlit-webrtc
av
numpy==1.24.4 
pandas==1.5.3
